+++ cosmohub/api/sockets/queries.py
@@ -1,7 +1,9 @@
-import gevent
 import json
+import logging
+import time
 
 from flask import g, current_app
+from geventwebsocket import WebSocketError
 
 from cosmohub.api import db, ws
 
@@ -10,8 +12,12 @@ from ..db.session import transactional_session
 from ..security.authentication import verify_token
 from ..utils import webhcat
 
+log = logging.getLogger(__name__)
+
 @ws.route('/sockets/queries')
 def echo_socket(ws):
+    log.info("Opened websocket connection")
+    
     hive_rest = webhcat.Hive(
         url = current_app.config['WEBHCAT_BASE_URL'],
         username = 'jcarrete',
@@ -24,7 +30,9 @@ def echo_socket(ws):
         # Do not proceed if there is not valid token
         if not msg['type'] == 'auth' or not verify_token(msg['data']['token']):
             return
-    
+        
+        old_set = set([])
+        new_set = set([])
         while not ws.closed:
             with transactional_session(db.session) as session:
                 queries = session.query(model.Query).filter_by(
@@ -34,29 +42,37 @@ def echo_socket(ws):
                     model.Query.ts_submitted
                 )
             
+            data = {}
+            new_set.clear()
             for query in queries:
+                new_set.add(query.id)
+                old_set.remove(query.id)
+                
                 status = hive_rest.status(query.job_id)
                 progress = status["percentComplete"]
                 percent = int(progress[:progress.index('%')])
                 
-                ws.send(json.dumps({
-                    'type' : 'progress',
-                    'data' : {
-                        query.id : percent,
-                    }
-                }))
+                data[query.id] = percent
             
-            # FIXME: Remove
-            import random
+            if old_set:
+                # Some tracked query has completed.
+                # Send non existing query ID to force refresh
+                data[0] = 0
+
+            # Send query progress
             ws.send(json.dumps({
                 'type' : 'progress',
-                'data' : {
-                    42 : random.randint(0, 100),
-                }
+                'data' : data,
             }))
             
-            gevent.sleep(1)
+            old_set = new_set
+            
+            time.sleep(5)
     
+    except WebSocketError:
+        pass
     except TypeError:
         if not ws.closed:
             raise
+
+    log.info("Closing websocket connection")